Tests for pole depths
